+++ NEWS.md
@@ -7,6 +7,9 @@
 - added dig_tautologies()
 - added geom_diamond()
 - added .span and .inc arguments to partition()
+- added various styles (quantile, kmeans, hclust, bclust, ...) of crisp
+  partitioning to partition()
+- added explore() function for interactive exploration of patterns
 - added exclude argument to dig() and other dig_*() functions
 - added support for the disjoint parameter to var_grid() and dig_grid()
 - added progress bar to the dig() function